file: Adding analyse_heat.sh to repository
The script analyse_heat.sh passes on all resources that failed
to be created/updated (via heat), thus extracting the stdout/stderr
and linking the messages with the correct server name/IP address.

Basically it avoids typing always the same command, and gives a faster
overview on what failed and where.